pyproject.toml
[project]
name = "my-agent"
version = "0.1.0"
description = "Content creation AI agents powered by CrewAI"
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=1.10.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"
